Update README.md
Heart stroke prediction refers to the assessment of an individual's risk of experiencing a stroke, specifically related to cardiovascular health. A stroke occurs when the blood supply to the brain is disrupted, leading to the death of brain cells due to a lack of oxygen and nutrients. It is a serious medical condition that can have long-lasting consequences or even be fatal.